+++ modules/raft/src/test/java/org/apache/ignite/raft/jraft/rpc/AbstractRpcTest.java
@@ -16,12 +16,6 @@
  */
 package org.apache.ignite.raft.jraft.rpc;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Queue;
@@ -34,6 +28,12 @@ import org.apache.ignite.raft.messages.TestRaftMessagesFactory;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.fail;
 
 /**
  *
@@ -97,18 +97,18 @@ public abstract class AbstractRpcTest {
         RpcClient client = createClient();
 
         CountDownLatch l1 = new CountDownLatch(1);
-        AtomicReference<Response1> resp1 = new AtomicReference<>();
+        AtomicReference<TestMessages.Response1> resp1 = new AtomicReference<>();
         client.invokeAsync(peerId, msgFactory.request1().build(), new InvokeContext(), (result, err) -> {
-            resp1.set((Response1) result);
+            resp1.set((TestMessages.Response1) result);
             l1.countDown();
         }, 5000);
         l1.await(5_000, TimeUnit.MILLISECONDS);
         assertNotNull(resp1);
 
         CountDownLatch l2 = new CountDownLatch(1);
-        AtomicReference<Response2> resp2 = new AtomicReference<>();
+        AtomicReference<TestMessages.Response2> resp2 = new AtomicReference<>();
         client.invokeAsync(peerId, msgFactory.request2().build(), new InvokeContext(), (result, err) -> {
-            resp2.set((Response2) result);
+            resp2.set((TestMessages.Response2) result);
             l2.countDown();
         }, 5000);
         l2.await(5_000, TimeUnit.MILLISECONDS);
@@ -159,7 +159,7 @@ public abstract class AbstractRpcTest {
         assertTrue(client1.checkConnection(peerId));
 
         try {
-            Request1 request = msgFactory.request1().val(10_000).build();
+            TestMessages.Request1 request = msgFactory.request1().val(10_000).build();
 
             CompletableFuture<Object> fut = new CompletableFuture<>();
 
@@ -181,13 +181,13 @@ public abstract class AbstractRpcTest {
         Queue<Object[]> recorded = client1.recordedMessages();
 
         assertEquals(1, recorded.size());
-        assertTrue(recorded.poll()[0] instanceof Request1);
+        assertTrue(recorded.poll()[0] instanceof TestMessages.Request1);
     }
 
     @Test
     public void testBlockedAsync() throws Exception {
         RpcClientEx client1 = (RpcClientEx) createClient();
-        client1.blockMessages((msg, id) -> msg instanceof Request1);
+        client1.blockMessages((msg, id) -> msg instanceof TestMessages.Request1);
 
         assertTrue(client1.checkConnection(peerId));
 
@@ -209,9 +209,9 @@ public abstract class AbstractRpcTest {
     }
 
     /** */
-    private class Request1RpcProcessor implements RpcProcessor<Request1> {
+    private class Request1RpcProcessor implements RpcProcessor<TestMessages.Request1> {
         /** {@inheritDoc} */
-        @Override public void handleRequest(RpcContext rpcCtx, Request1 request) {
+        @Override public void handleRequest(RpcContext rpcCtx, TestMessages.Request1 request) {
             if (request.val() == 10_000)
                 try {
                     Thread.sleep(1000);
@@ -220,27 +220,27 @@ public abstract class AbstractRpcTest {
                     // No-op.
                 }
 
-            Response1 resp1 = msgFactory.response1().val(request.val() + 1).build();
+            TestMessages.Response1 resp1 = msgFactory.response1().val(request.val() + 1).build();
             rpcCtx.sendResponse(resp1);
         }
 
         /** {@inheritDoc} */
         @Override public String interest() {
-            return Request1.class.getName();
+            return TestMessages.Request1.class.getName();
         }
     }
 
     /** */
-    private class Request2RpcProcessor implements RpcProcessor<Request2> {
+    private class Request2RpcProcessor implements RpcProcessor<TestMessages.Request2> {
         /** {@inheritDoc} */
-        @Override public void handleRequest(RpcContext rpcCtx, Request2 request) {
-            Response2 resp2 = msgFactory.response2().val(request.val() + 1).build();
+        @Override public void handleRequest(RpcContext rpcCtx, TestMessages.Request2 request) {
+            TestMessages.Response2 resp2 = msgFactory.response2().val(request.val() + 1).build();
             rpcCtx.sendResponse(resp2);
         }
 
         /** {@inheritDoc} */
         @Override public String interest() {
-            return Request2.class.getName();
+            return TestMessages.Request2.class.getName();
         }
     }
